Move Ul and H2 into shared Add border bottom to h2 and summary
