+++ src/components/Shared/H2.tsx
@@ -1,0 +1,8 @@
+import styled from "@emotion/styled";
+
+import colors from "../../styles/colors";
+
+export default styled.h2<{ hasBorderBottom?: boolean }>((props) => ({
+  paddingBottom: props.hasBorderBottom ? 8 : 0,
+  borderBottom: props.hasBorderBottom ? `1px solid ${colors.grey}` : 0,
+}));
