+++ src/components/Shared/Ul.tsx
@@ -1,0 +1,8 @@
+import styled from "@emotion/styled";
+
+export default styled.ul<{ gap: number }>((props) => ({
+  marginLeft: "1.25rem",
+  "& > li": {
+    marginTop: props.gap,
+  },
+}));
